Rework templatetag civ rendering into pure templates
