+++ app/grandchallenge/components/templates/components/partials/civ/file_thumbnail.html
