+++ app/grandchallenge/components/templates/components/partials/civ/value.html
