+++ app/grandchallenge/components/templates/components/partials/civ/value_chart.html
