+++ app/grandchallenge/components/templates/components/partials/civ_inline.html
@@ -1,0 +1,30 @@
+{% load civ %}
+
+{% with interface=object.interface civ=object %}
+
+    {% if interface.super_kind == interface.SuperKind.VALUE %}
+        {# No fallback here as a value could be null #}
+        {% if interface.kind == interface.Kind.CHART %}
+            {% include "components/partials/civ_inline/value_chart.html" %}
+        {% elif civ|can_preview %}
+            {% include "components/partials/civ_inline/value_preview.html" %}
+        {% else %}
+            {% include "components/partials/civ_inline/value.html" %}
+        {% endif %}
+    {% elif interface.super_kind == interface.SuperKind.FILE %}
+        {% if not civ.file %}
+            {% include "components/partials/civ_inline/fallback.html" %}
+        {% elif civ.is_thumbnail_kind %}
+            {% include "components/partials/civ_inline/file_thumbnail.html" %}
+        {% else %}
+            {% include "components/partials/civ_inline/file.html" %}
+        {% endif %}
+    {% elif interface.super_kind == interface.SuperKind.IMAGE %}
+        {% if not civ.image %}
+            {% include "components/partials/civ_inline/fallback.html" %}
+        {% else %}
+            {% include "components/partials/civ_inline/image.html" %}
+        {% endif %}
+    {% endif %}
+
+{% endwith %}
