+++ app/grandchallenge/components/templates/components/partials/civ_inline/fallback.html
@@ -1,0 +1,1 @@
+{% include "components/partials/civ/fallback.html" %}
