+++ app/grandchallenge/components/templates/components/partials/civ_inline/file.html
@@ -1,4 +1,4 @@
-{% extends "../file.html" %}
+{% extends "components/partials/civ/file.html" %}
 
 {% load pathlib %}
