+++ app/grandchallenge/components/templates/components/partials/civ_inline/file_thumbnail.html
@@ -1,0 +1,5 @@
+{% extends "components/partials/civ_inline/modal_base.html" %}
+
+{% block content %}
+    {% include "components/partials/civ//file_thumbnail.html" %}
+{% endblock %}
