+++ app/grandchallenge/components/templates/components/partials/civ_inline/image.html
@@ -1,4 +1,4 @@
-{% extends "../image.html" %}
+{% extends "components/partials/civ/image.html" %}
 
 {% load pathlib %}
