+++ app/grandchallenge/components/templates/components/partials/civ_inline/modal_base.html
