+++ app/grandchallenge/components/templates/components/partials/civ_inline/value.html
@@ -1,4 +1,4 @@
-{% extends "./modal_base.html" %}
+{% extends "components/partials/civ_inline/modal_base.html" %}
 
 {% load json %}
