+++ app/grandchallenge/components/templates/components/partials/civ_inline/value_chart.html
@@ -1,4 +1,4 @@
-{% extends "./modal_base.html" %}
+{% extends "components/partials/civ_inline/modal_base.html" %}
 
 {% block content %}
