+++ app/grandchallenge/components/templates/components/partials/civ_inline/value_preview.html
@@ -1,4 +1,4 @@
-{% extends "./value.html" %}
+{% extends "components/partials/civ_inline/value.html" %}
 
 {% load json %}
