--- app/grandchallenge/core/templates/grandchallenge/partials/component_interface_values/inline/fallback.html
@@ -1,1 +1,0 @@
-{% include "../fallback.html" %}
