--- app/grandchallenge/core/templates/grandchallenge/partials/component_interface_values/inline/file_thumbnail.html
@@ -1,5 +1,0 @@
-{% extends "./modal_base.html" %}
-
-{% block content %}
-    {% include "../file_thumbnail.html" %}
-{% endblock %}
